backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymupdf
chromadb
openai
google-generativeai
tiktoken
pydantic-settings
python-multipart